requirements.txt
anyio==4.7.0
beautifulsoup4==4.12.3
certifi==2024.12.14
click==8.1.7
colorama==0.4.6
fastcore==1.7.27
fastlite==0.0.13
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
peewee==3.17.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.10.1
python-multipart==0.0.20
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlite-minutils==4.0.3
starlette==0.42.0
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
